Squashing warnings

+++ GAM200/src/Assets Manager/asset_manager.cpp
@@ -299,7 +299,7 @@ FMOD::Sound* AssetManager::getsoundbyaudiotype(AudioType a, bool random, int seq
 				std::random_device rd;
 				std::mt19937 gen(rd());
 
-				std::uniform_int_distribution<> dis(0, v.size() - 1);
+				std::uniform_int_distribution<> dis(0, (int) v.size() - 1);
 
 				return soundsval(v[dis(gen)]);
 			}
